% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_pprof_pkg_path}
\alias{get_pprof_pkg_path}
\title{Return the path to the pprof executable}
\usage{
get_pprof_pkg_path()
}
\description{
It remains the caller's responsibility to run the process.
}
\examples{
system2(get_pprof_pkg_path())
}
