% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprof-package.R
\docType{package}
\name{pprof-package}
\alias{pprof}
\alias{pprof-package}
\title{pprof: Installation Helper for the 'pprof' Tool}
\description{
Contains sources and a build script for the 'pprof' tool by 'Google'.
}
\details{
The \code{pprof} binary can also be installed via \code{go get github.com/google/pprof}.
However, some operating systems have an incompatible version of \code{pprof} on
the \code{PATH}. The R package offers an easy-to-use robust alternative.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-prof/r-pprof}
  \item \url{http://r-prof.github.io/r-pprof}
  \item Report bugs at \url{https://github.com/r-prof/r-pprof/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Other contributors:
\itemize{
  \item R Consortium [funder]
  \item Google, Inc. (Bundled pprof sources) [copyright holder]
}

}
